refactor loginPage to LoginForm
