+++ client/src/components/forms/LoginForm.tsx
@@ -1,0 +1,65 @@
+import { Alert, Box, IconButton, Stack, TextField, TextFieldProps } from '@mui/material';
+import CloseIcon from '@mui/icons-material/Close';
+import { AxiosResponse } from 'axios';
+import MainButton from '../UI/buttons/MainButton';
+import { columnCenter } from '../../styles/flex';
+import LoadingButton from '../UI/buttons/LoadingButton';
+import { Nullable } from '../../types/utils/Nullable';
+
+interface Props {
+  error: Nullable<AxiosResponse>;
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  setError: React.Dispatch<React.SetStateAction<Nullable<AxiosResponse<any, any>>>>;
+  isLoading: boolean;
+  onSubmit: (e: React.FormEvent<HTMLFormElement>) => Promise<void>;
+  firstFieldProps: TextFieldProps;
+  secondFieldProps: TextFieldProps;
+  submitButtonText: string;
+}
+
+const LoginForm: React.FC<Props> = ({
+  error,
+  setError,
+  onSubmit,
+  isLoading,
+  firstFieldProps,
+  secondFieldProps,
+  submitButtonText,
+}) => {
+  return (
+    <Box component="form" noValidate sx={{ gap: '20px', ...columnCenter }} onSubmit={onSubmit}>
+      <Stack width="300px" direction="column" spacing={2}>
+        <TextField disabled={isLoading} {...firstFieldProps} />
+        <TextField disabled={isLoading} {...secondFieldProps} />
+      </Stack>
+      <LoadingButton
+        buttonBase={MainButton}
+        loading={isLoading}
+        disableElevation
+        variant="contained"
+        type="submit"
+      >
+        {submitButtonText}
+      </LoadingButton>
+      {error && (
+        <Alert
+          severity="error"
+          action={
+            <IconButton
+              size="small"
+              aria-label="close"
+              color="inherit"
+              onClick={() => setError(null)}
+            >
+              <CloseIcon fontSize="small" />
+            </IconButton>
+          }
+        >
+          {error.data.statusCode === 401 ? 'Wrong email or password' : 'Server error'}
+        </Alert>
+      )}
+    </Box>
+  );
+};
+
+export default LoginForm;
